backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
openai==1.12.0
qdrant-client==1.7.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
langchain==0.1.6
langchain-openai==0.0.5
pydantic==2.6.1
python-multipart==0.0.9